requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-learn>=1.3.0
albumentations>=1.3.0
mtcnn>=0.1.1
tensorboard>=2.13.0
kagglehub>=0.2.0
jupyter>=1.0.0
ipywidgets>=8.0.0

